+++ cmd/neofs-cli/modules/accounting/balance.go
@@ -1,7 +1,6 @@
 package accounting
 
 import (
-	"context"
 	"fmt"
 	"math/big"
 
@@ -40,7 +39,9 @@ var accountingBalanceCmd = &cobra.Command{
 			return fmt.Errorf("can't decode owner ID wallet address: %w", idUser.DecodeString(balanceOwner))
 		}
 
-		ctx := context.Background()
+		ctx, cancel := commonflags.GetCommandContext(cmd)
+		defer cancel()
+
 		cli, err := internalclient.GetSDKClientByFlag(ctx, commonflags.RPC)
 		if err != nil {
 			return err
@@ -62,11 +63,10 @@ var accountingBalanceCmd = &cobra.Command{
 }
 
 func initAccountingBalanceCmd() {
+	commonflags.Init(accountingBalanceCmd)
+
 	ff := accountingBalanceCmd.Flags()
 
-	ff.StringP(commonflags.WalletPath, commonflags.WalletPathShorthand, commonflags.WalletPathDefault, commonflags.WalletPathUsage)
-	ff.StringP(commonflags.Account, commonflags.AccountShorthand, commonflags.AccountDefault, commonflags.AccountUsage)
-	ff.StringP(commonflags.RPC, commonflags.RPCShorthand, commonflags.RPCDefault, commonflags.RPCUsage)
 	ff.String(ownerFlag, "", "owner of balance account (omit to use owner from private key)")
 }